Convert from initializer to generator.
